requirements.txt
autopep8==1.6.0
bcrypt==3.2.0
black==21.12b0
blinker==1.4
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.0.0
platformdirs==2.4.1
psycopg2==2.9.3
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
python-dotenv==0.19.2
python-jose==3.3.0
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.29
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
Werkzeug==2.0.2
WTForms==3.0.1
yapf==0.32.0
